Updates clean rules #17
The descriptor files are included in the file list of descriptors
because of missing clean rules: the contents of the descriptor is
generated from a file list, that may contain files from old builds. This
change ensures old descriptors are removed explicitly before calculating
the file list for the new descriptor files.

This change ensures that `gradle clean` removes everything except the
extracted MagicDraw instance used as jar dependency source, thus further
reducing the potential of files of old builds to show up.