Simplifies model set updating code in engine context #37
Before V4MD 2.2.1 navigationHelper#getModelRoots() returned the internal
set of model roots; starting with 2.2.1 only a current snapshot is
returned. This broke the traversal functionality in
MagicDrawProjectEngineContext#modelSetUpdated(), as the difference set
never became empty. On the other hand, it became possible to simply
traverse the rootsToRemove and rootsToAdd sets without any specific
checking, as the lack of connections between the internal structures of
the NavigationHelper means there is no chance for
ConcurrentModificationExceptions during traversal.