Ensures IncrementalTransformationSynchronizer cleanup on project close
When a project is closed where an IncrementalTransformationSynchronizer
was initialized, and for an error a reference to the project was kept,
the WeakHashMap entry for the synchronizers were not removed, resulting
in the query engine being kept in memory.

This change ensures that even in such erroneous cases the engine is
disposed by explicitly calling a remove call before a project is closed.